+++ src/ezdxf/addons/__init__.py
@@ -1,6 +1,6 @@
 # Copyright (C) 2011-2022, Manfred Moitzi
 # License: MIT License
-from .table import Table, CustomCell
+from .tablepainter import TablePainter, CustomCell
 from .mtext import MTextSurrogate
 from .menger_sponge import MengerSponge
 from .sierpinski_pyramid import SierpinskyPyramid